requirements.txt
Django>=3.2.25
djangorestframework>=3.15.1
django-cors-headers>=4.3.1
python-decouple>=3.8
whitenoise>=6.6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
